added support for meshes with additional assets
in case someone saves a body mesh and adds assets (e.g. hair, fingernails, eyes, tongue) to this body one could also work on the character and use a formerly created symmetry file.

An example e.g. is a makehuman character which gets additional shape-keys for Unreal Engine. It usually is equipped with eyes etc and all the methods will no longer work.

the additional vertices are ignored and a warning will appear. It works on
vertex groups, mesh and shapekeys. Better than creating a stack dump ;)